feat: refactored Tabs: save/restore now in site.store; hmr works
